pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sage-multi-tenancy"
version = "0.1.0"
description = "Gateway, scheduler, and BioCLIP torch worker for Sage multi-tenant inference on Orin"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.26.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.6",
  "pywaggle[all]",
]

[project.optional-dependencies]
torch-worker = [
  "torch",
  "open-clip-torch",
  "Pillow>=10.0.0",
]

[project.scripts]
sage-gateway = "sage_mt.gateway.app:main"
sage-torch-worker = "sage_mt.torch_worker.app:main"
sage-ui = "sage_mt.web_ui.app:main"

[tool.setuptools.packages.find]
where = ["src"]
